refactor hooks module
