+++ client/src/App.css
@@ -87,7 +87,7 @@ header {
   align-items: center;
   width: 100vw;
   height: 100vh;
-  font-family: var(--mono);
+  font-family: var(--font-primary);
 }
 
 .registration-form > div > input {
@@ -169,13 +169,13 @@ header {
 }
 
 .messages-container {
-  font-size: 1rem;
+  font-size: var(--base-font-size);
   word-break: break-all;
   text-align: left;
   width: 90%;
   padding-left: 3rem;
   padding-right: 3rem;
-  font-family: var(--mono);
+  font-family: var(--font-primary);
 }
 
 .message-username::before {
@@ -192,7 +192,7 @@ header {
   border-top: 1px solid rgba(20, 20, 20, 0.9);
   padding: 0.66rem 0.33rem;
   border-radius: 0 7% 0 0;
-  margin: 1rem;
+  margin: 1rem 1rem 0.33rem;
   box-shadow: 5px -5px 21px 1px rgba(25, 25, 25, 0.01);
   border-radius: 6px;
   word-break: keep-all;
@@ -207,7 +207,7 @@ header {
 
 .message-username,
 .datetime {
-  font-size: 0.66rem;
+  font-size: 0.66em;
 }
 
 .datetime {
@@ -220,7 +220,6 @@ header {
 
 .message-text {
   padding-left: 2rem;
-  padding-bottom: 1rem;
 }
 
 .name-and-date {
