+++ client/src/components/ChatView.tsx
@@ -1,18 +1,16 @@
-import React, { useEffect, useState } from "react";
+import React, { Suspense, useEffect, useState } from "react";
 import Messages, { MessageArray } from "./Messages";
 import ChatForm from "./Forms/ChatForm";
 import SessionUtils from "./SessionUtils";
 import { API } from "../api/api";
 
 const ChatView: React.FC<any> = ({ socket, currentRoom }) => {
-  const [loading, setLoading] = useState(true);
   const [localMessageState, setLocalMessageState] = useState<MessageArray>([]);
 
   useEffect(() => {
     const getMessages = async () => {
       const messages = await API.getRecentMessages(currentRoom);
       setLocalMessageState(messages.data.reverse());
-      setLoading(false);
     };
 
     getMessages();
@@ -20,19 +18,15 @@ const ChatView: React.FC<any> = ({ socket, currentRoom }) => {
 
   return (
     <>
-      {loading ? (
-        <div className="messages-loading">Fetching messages ... </div>
-      ) : (
-        <>
-          <Messages
-            socket={socket}
-            messages={localMessageState}
-            setMessageState={setLocalMessageState}
-          />
-          <ChatForm socket={socket} currentRoom={currentRoom} />
-          <SessionUtils />
-        </>
-      )}
+      <Suspense fallback={<>Loading ...</>}>
+        <Messages
+          socket={socket}
+          messages={localMessageState}
+          setMessageState={setLocalMessageState}
+        />
+        <ChatForm socket={socket} currentRoom={currentRoom} />
+        <SessionUtils />
+      </Suspense>
     </>
   );
 };
