+++ client/src/index.css
@@ -1,5 +1,7 @@
 @import url("https://fonts.googleapis.com/css2?family=Anonymous+Pro&display=swap");
+@import url("https://fonts.googleapis.com/css2?family=Dongle:wght@300&display=swap");
 @import url("./App.css");
+
 :root {
   font-family: "Anonymous Pro", monospace;
   line-height: 1.5;
@@ -94,9 +96,7 @@ button:hover {
   .message-content {
     border-right: unset;
   }
-  .datetime {
-    font-size: 0.5rem;
-  }
+
   .app-title {
     margin-top: 1rem;
   }
